Add files via upload
I changed the scipy.optimize.curve_fit function absolute_sigma flag to true to indicate that the errors in sigma are treated in the absolute sense, allowing the error reported in the pcov matrix to correspond to the error in the fit parameters. Otherwise, only the relative magnitude of sigma is taken into account and the pcov matrix is based on scaling sigma by a constant factor. I also made it print the calculated error in the fits below where it prints the fit parameters.